backend/requirements.txt
aioredis==1.3.1
anyio==3.6.2
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.3.2
Automat==20.2.0
boto3==1.22.2
botocore==1.25.2
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.1
daphne==3.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.0.4
django-anymail==8.6
django-cors-headers==3.11.0
django-debug-toolbar==3.5.0
django-rest-passwordreset==1.2.1
django-storages==1.12.3
django-templated-mail==1.1.1
django-urls==1.1.3
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
gunicorn==20.1.0
h11==0.14.0
hiredis==2.0.0
httptools==0.5.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.0
MarkupSafe==2.1.1
msgpack==1.0.3
oauthlib==3.2.0
pendulum==2.1.2
Pillow==9.1.0
psycopg==3.0.11
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.1
pytzdata==2020.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
s3transfer==0.5.2
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
stripe==2.74.0

txaio==22.2.1
typing_extensions==4.2.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
uvicorn==0.19.0
watchfiles==0.18.1
websockets==10.4
zope.interface==5.4.0
